requirements.txt
web3>=7.0.0
anthropic>=0.40.0
eth-account>=0.13.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
flask>=3.0.0
